+++ server/tests/README.md
@@ -1,0 +1,1 @@
+## run test --models=test
